Invoice due date is getting erased when payment process is failed
Invoice due date was getting erased when a user fails to complete the payment process.
For example, a user starts to pay an invoice with Stripe. The Stripe overlay is displayed,
however the user decides not to pay at this time.
In this case the due date would have been erased. This will correct the issue.